chore: Added `ProcessedCli` to extract the logic to pre-process CLI struct args
