+++ crates/intrinsic-test/src/common/supporting_test.rs
@@ -1,13 +1,10 @@
+use crate::common::cli::ProcessedCli;
+
 /// Architectures must support this trait
 /// to be successfully tested.
 pub trait SupportedArchitectureTest {
-    fn write_c_file(filename: &str);
-
-    fn write_rust_file(filename: &str);
-
-    fn build_c_file(filename: &str);
-
-    fn build_rust_file(filename: &str);
-
-    fn read_intrinsic_source_file(filename: &str);
+    fn create(cli_options: ProcessedCli) -> Self;
+    fn build_c_file(&self) -> bool;
+    fn build_rust_file(&self) -> bool;
+    fn compare_outputs(&self) -> bool;
 }
